Небольшой багфикс

+++ src/model/RelationData.h
@@ -6,6 +6,7 @@
 
 #include "RelationalSchema.h"
 #include "ColumnData.h"
+#include <iostream>
 #include <vector>
 
 using std::vector;
@@ -22,7 +23,9 @@ public:
     virtual vector<int> getTuple(int tupleIndex)=0;
     virtual void shuffleColumns()=0;
     double getMaximumNip() { return getNumRows() * (getNumRows() - 1) / 2.0; }
-    unsigned int getNumTuplePairs() { return getNumRows() * (getNumRows() - 1) / 2; }
+    long long getNumTuplePairs() {
+        //std::cout << "----" << (long long) getNumRows() * (getNumRows() - 1) / 2 << std::endl;
+        return (long long) getNumRows() * (getNumRows() - 1) / 2; }
 
     shared_ptr<RelationalSchema> getSchema();
 protected:
